Extract report sub-package
