+++ src/mireport/report/inline_report_templates/inline-report-flat-list.html.jinja
