+++ src/mireport/report/inline_report_templates/style-efrag.css.jinja
