+++ src/mireport/report/inline_report_templates/style.css.jinja
